shiftfs_lookup(): new->d_inode is not stable
BugLink: https://bugs.launchpad.net/bugs/

The lower dentry can't go from positive to negative (we have it pinned),
but it *can* go from negative to positive.  So fetching ->d_inode into a
local variable, doing a blocking allocation, checking that now ->d_inode
is non-NULL and feeding the value we'd fetched earlier to a function
that won't accept NULL is not a good idea.

Link: 5873f1425234 ("ecryptfs_lookup_interpose(): lower_dentry->d_inode is not stable")
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>